requirements.txt
fastapi==0.95.1
pydantic==1.10.2
numpy==1.24.3
pillow==10.0.1
scikit-learn==1.5.2
pandas==2.2.3
uvicorn==0.23.2
